=> Fixed a bug in the simple auction test where the 3 seconds wait was after the request => Verified the balances after the simple auction => Implemented a mechanism for making transfers between addresses that do not include the transaction sender => Implemented a more performant balance mechanism => Added a unit test for testing keccak256 array hash => Started implementing the balances as unint256_t as per the EVM spec => Added a utility method for formatting the EVM uint256_t balance as an EOS asset => Fixed the proposal tests to ensure that both approvals are required before the proposal is marked as approved => Fixed a bug where RLP encoded signatures that contained leading 0's in the R or S component were not being resolved. This issue is covered by the MalformedTransactionDebugTest which verifies this has been fixed against 10,000 signatures. => Fixed the WEI test to include the EOS asset 4 precision

+++ integration_tests/src/main/kotlin/com/memtrip/eos_evm/eos/evm/contracts/misc/MiniDAOContract.kt
@@ -44,10 +44,10 @@ class MiniDAOContract(
         )
     }
 
-    fun withdraw(amount: Long, sender: EvmSender): Single<ChainResponse<TransactionCommitted>> {
+    fun withdraw(amount: BigInteger, sender: EvmSender): Single<ChainResponse<TransactionCommitted>> {
         return executeMethod(
             "withdraw",
-            listOf(Uint(BigInteger.valueOf(amount))),
+            listOf(Uint(amount)),
             listOf(),
             sender
         )
